Add SIMD-based JSON iterator benchmarks, extend `json_simd_iterator` with class and boolean iterator implementations, and optimize grammar validation functions.
